refactor: remove worker kind from repository
