+++ server/libs/common/src/models/workers.rs
@@ -4,7 +4,7 @@ use utoipa::ToSchema;
 
 use time::OffsetDateTime;
 
-use super::TaskKind;
+use super::{TaskInstance, TaskKind};
 
 /// A worker that can execute tasks after receiving them.
 /// We know that it can receive those tasks from its list of capabilities.
@@ -31,5 +31,47 @@ impl Worker {
             task_kind,
             active: true,
         }
+    }
+
+    pub fn can_handle(&self, task: &TaskInstance) -> bool {
+        self.task_kind
+            .iter()
+            .any(|kind| kind.name == task.task_kind.name)
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use sqlx::types::Uuid;
+    use time::OffsetDateTime;
+
+    #[test]
+    fn test_worker_can_handle() {
+        let task_kind1 = TaskKind::new("task1".to_string());
+        let task_kind2 = TaskKind::new("task2".to_string());
+        let worker_kind = WorkerKind::new(vec![task_kind1.clone()]);
+
+        let task1 = TaskInstance {
+            id: Uuid::new_v4(),
+            task_kind: task_kind1.clone(),
+            status: crate::TaskStatus::Queued,
+            created_at: OffsetDateTime::now_utc(),
+            input_data: None,
+            assigned_to: None,
+            result: None,
+        };
+        let task2 = TaskInstance {
+            id: Uuid::new_v4(),
+            task_kind: task_kind2.clone(),
+            status: crate::TaskStatus::Queued,
+            created_at: OffsetDateTime::now_utc(),
+            input_data: None,
+            assigned_to: None,
+            result: None,
+        };
+
+        assert!(worker_kind.can_handle(&task1));
+        assert!(!worker_kind.can_handle(&task2));
     }
 }
